requirements.txt
# Flask Backend
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code Quality
pylint==3.0.3
black==23.12.1

# Optional: Redis for caching
# redis==5.0.1
